ci: add clippy + rustfmt gate, and make the tree clippy-clean (#162)
CI previously ran only `cargo test`. Add a `lint` job running
`cargo fmt --all --check` and `cargo clippy --workspace --all-targets
-- -D warnings` (installing protobuf-compiler/libsystemd-dev/libssl-dev so
every crate, incl. gnmi/snmp/logs, compiles under clippy).

`cargo clippy` did not even pass before this: deny-by-default
`approx_constant` fired on `3.14`/`3.14159` test fixtures. Fix the whole
backlog so the new gate is green:

- approx_constant: replace π-looking test fixtures with neutral values
  (2.5, or 1.23456 where a rounding assertion is the point).
- unnecessary_sort_by: descending sorts -> sort_by_key(Reverse(..))
  (netflow/syslog/netlink views, snmp mib).
- items_after_test_module: move netlink command::run above its #[cfg(test)].
- unnecessary_cast (modbus), inconsistent_digit_grouping (otel logs),
  needless is_none()->contains_key (netring), unused import (sensor-core),
  useless vec!/conversion/collapsible-if (zensight, via clippy --fix).
- field_reassign_with_default: struct literal where small (topology layout);
  scoped #[allow] for stepwise test-fixture setup (ui_tests, tokens).

+++ zensight-sensor-modbus/src/poller.rs
@@ -430,7 +430,7 @@ mod tests {
 
     #[test]
     fn test_decode_with_scale_offset() {
-        let raw = 1000u16 as f64;
+        let raw = 1000.0_f64;
         let scale = 0.1;
         let offset = -50.0;
         let result = raw * scale + offset;
